Big commit. Restructued all of preprocess to prepare to add opus dataset (and potentially others). Changed train/test code to match but testing to be done first. Added argument to exclude translations from training multilingual models with ted-multilingual.
