+++ test.py
@@ -183,10 +183,9 @@ def multi_test(device, params, test_dataloader, tokenizer, verbose=50):
     model, _, _, _ = logging.load_checkpoint(logger.checkpoint_path, device, model)
 
     assert tokenizer is not None
-    add_targets = preprocess.AddTargetTokens(params.langs, tokenizer)
-    pair_accs = {s+'-'+t : 0.0 for s, t in get_pairs(params.langs)}
+    pair_accs = {s+'-'+t : 0.0 for s, t in get_pairs(params.langs, excluded=params.excluded)}
     pair_bleus = {}
-    for s, t in get_pairs(params.langs):
+    for s, t in get_pairs(params.langs, excluded=params.excluded):
         _bleu = BLEU()
         _bleu.set_excluded_indices([0, 2])
         pair_bleus[s+'-'+t] = _bleu
@@ -198,10 +197,7 @@ def multi_test(device, params, test_dataloader, tokenizer, verbose=50):
 
     print("Now testing")
     for i, data in enumerate(test_dataloader):
-
-        data = get_directions(data, params.langs)
         for direction, (x, y, y_lang) in data.items():
-            x = add_targets(x, y_lang)
             bleu = pair_bleus[direction]
             test_batch_acc = inference_step(x, y, model, logger, tokenizer, device, bleu=bleu,
                                         teacher_forcing=params.teacher_forcing,
@@ -285,8 +281,8 @@ def main(params):
             tokenizers = None
 
         train_dataloader, val_dataloader, test_dataloader, tokenizers = preprocess.load_and_preprocess(
-            params.langs, params.batch_size, params.vocab_size, params.dataset,
-            tokenizer=tokenizers, multi=False)
+            params.dataset, params.langs, params.vocab_size, batch_size=params.batch_size, mode='bilingual',
+            pivot_pair_ind=params.pivot_pair_ind, tokenizer=tokenizers, excluded=params.excluded)
 
         test(device, params, test_dataloader, tokenizers, verbose=params.verbose)
 
@@ -299,8 +295,8 @@ def main(params):
             tokenizer = None
 
         train_dataloader, val_dataloader, test_dataloader, tokenizer = preprocess.load_and_preprocess(
-            params.langs, params.batch_size, params.vocab_size, params.dataset,
-            tokenizer=tokenizer, multi=True)
+            params.dataset, params.langs, params.vocab_size, batch_size=params.batch_size, mode='multi',
+            pivot_pair_ind=params.pivot_pair_ind, tokenizer=tokenizer, excluded=params.excluded)
 
         multi_test(device, params, test_dataloader, tokenizer, verbose=params.verbose)
 
@@ -311,8 +307,8 @@ def main(params):
             tokenizer = None
 
         train_dataloader, val_dataloader, test_dataloader, tokenizer = preprocess.load_and_preprocess(
-            params.langs, params.batch_size, params.vocab_size, params.dataset,
-            tokenizer=tokenizer, multi=True)
+            params.dataset, params.langs, params.vocab_size, batch_size=params.batch_size, mode='pivot',
+            pivot_pair_ind=params.pivot_pair_ind, tokenizer=tokenizer, excluded=params.excluded)
 
         pivot_test(device, params, test_dataloader, tokenizer, verbose=params.verbose)
